+++ src/contracts/index.js
@@ -1,8 +1,8 @@
-import ethersjs from 'ethers'
+import * as ethersjs from 'ethers'
 import registryAbi from './abi/registry.abi.js'
 import { strToBytes32, isEmpty } from '../utils/helpers.js'
 
-const { ethers } = ethersjs
+const { ethers } = ethersjs.default || ethersjs
 
 export const createProvider = providerUri => {
   if (isEmpty(providerUri)) throw Error('Empty web3 provider uri')